refactor: ♻️ reorganización de 'features/signup' y referencias actualizadas
- Archivos separados en carpetas según propósito en la feature 'signup'.
- Actualizadas referencias en 'app/(pages)/signup/page.tsx' para reflejar la nueva organización.
