+++ src/features/signup/hooks/useSignupContext.tsx
@@ -1,0 +1,11 @@
+import { useContext } from "react";
+import { SignupContextProps } from "../types/SignupContextProps";
+import { SignupContext } from "../contexts/SignupContext";
+
+export const useSignupContext = (): SignupContextProps => {
+  const context = useContext(SignupContext);
+  if (!context) {
+    throw new Error("useSignupContext must be used within a SignupProvider");
+  }
+  return context;
+};
