+++ src/features/signup/types/SignupFormData.ts
@@ -1,0 +1,6 @@
+export interface SignupFormData {
+  firstName: string;
+  lastName: string;
+  email: string;
+  password: string;
+}
